+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Updated rxdart to 0.25.0
 * Added working example/ folder
-* Added `isClosed` and `isOpen` method to all implementations
+* Added `isClosed` and `.add` method to all implementations
 
 ## [1.3.1] - 08/05/2020